moved components to components folder.
