+++ src/components/Home/style.scss
@@ -1,0 +1,34 @@
+@import '../../styles/variables.scss';
+
+.list{
+
+  .lastVisit{
+    padding-top: 50px;
+
+    .title{
+      padding: 25px 0px;
+    }
+  }
+}
+
+.item{
+  color: $brand-primary;
+
+  &:hover{
+    background: #f1f1f1;
+  }
+
+  .link{
+    display: block;
+    padding: 10px 10px;
+  }
+  .index{
+    opacity: 0.5;
+  }
+  .english{
+    font-size: 10px;
+  }
+  .arabic{
+    font-size: 14px;
+  }
+}
